+++ src/server.ts
@@ -1,12 +1,14 @@
 import express from 'express';
-import { configureOpenApi } from './zod-to-openapi';
+import { configureOpenApi } from './zod-openapi-express-routes/zod-to-openapi';
 import { productController } from './products/products.controller';
 
-const app = configureOpenApi({
-  app: express()
-    .use(express.json({ limit: '1mb' }))
-    .use(express.urlencoded({ extended: false, limit: '1mb' })),
-  routers: [productController],
+const expressApp = express()
+  .use(express.json({ limit: '1mb' }))
+  .use(express.urlencoded({ extended: false, limit: '1mb' }));
+
+const server = configureOpenApi({
+  app: expressApp,
+  controllers: [productController],
   docInfo: {
     title: 'ACME Products API',
     version: '0.1.3-dev',
@@ -16,4 +18,4 @@ const app = configureOpenApi({
   },
 });
 
-export default app;
+export default server;
